add solr search engine (partly fixes issue#35)
This commit adds solr search engine and indexes projects and users so
that we will be able to implement a search in the next commit.

In order to start solr engine, you need to run:

rake sunspot:solr:start

and then run

rake sunspot:reindex

for indexing the models in the database

Then, you can check it works entering into a rails console and typing:

> result = Project.search do
>     keywords "test"
>   end
> result.results

Assuming you have a project with the keyword "test" in either the title
or the description, you should get this/these project/s as a result.