add sunspot solr deploy tasks to mina
I've based this changes on the webistrano recipe that we have for other
projects.

I am adding start, stop, restart and reindex and I am forcing restart on
each deployment because deployment means changing directories and I
think that is the reason we need to restart sunspot.

Reindexing is only needed if you change the models or the searchable
blocks in the models and it is a very expensive operation, thus we
should not run it on every deploy.