Hide comment style bar in preview mode
Every comment field has a couple of style options, like adding emojis
and links.
The events that get triggered by pressing these buttons are attached
to the comment's textfield and break when the field is gone. Which is
the case in preview mode.
Hiding these buttons in preview mode solves the problem. And these
buttons are not of much use in preview mode anyway.