Fix travis
Travis was throwing the next exception:
`can't find gem bundler (>= 0.a) with executable bundle (Gem::GemNotFoundException)`
This is happing because RubyGems was unable to find the exact version of Bundler
that is in our Gemfile.lock. This bug is fixed in RubyGems 2.7.10, but
on Travis we are using 2.7.3. For more information
[Bundler Blog](https://bundler.io/blog/2019/05/14/solutions-for-cant-find-gem-bundler-with-executable-bundle.html).